requirements.txt
streamlit
xlwings
python-docx
pdf2image
Pillow
